Automatically update the "latest release" links
Today, Philip sent an email announcing the PMW 5.21 release, with the
previous release being the 5.20 release.

However, this page still listed PMW 5.10 as the "most recent" release,
which is inaccurate.

Rather than adding some code that auto-commits changes, or some github
automation to test, use some bits of Javascript to query the GitHub API
for the latest release, and then update the page with the correct
information.

As this temporarily shows the "old" information until the API request
has happened and was handled, replace the old 5.10 information with a
placeholder.